requirements.txt
torch>=2.0
transformers
tokenizers
datasets
evaluate
huggingface_hub
wandb
black
coloredlogs
transliterate
pyguitarpro